feat: Make mutations accessible
