+++ libs/index.js
@@ -8,5 +8,6 @@ libs.ACH = require('./ACH')
 libs.askConfirmation = require('./askConfirmation')
 libs.log = require('./log')
 libs.mkAPI = require('./api')
+libs.mutations = require('./mutations')
 
 module.exports = libs
